[AIRFLOW-6531] Move base classes to init to test "test_providers_modules_should_have_tests"
